+++ language/src/main/java/i4gl/nodes/statement/DisplayNode.java
@@ -15,6 +15,11 @@ public abstract class DisplayNode extends StatementNode {
     protected abstract ExpressionNode getArgumentNode();
 
     @Specialization
+    protected void display(final short argument) {
+        Context.get(this).getOutput().println(String.format("%11d", argument));
+    }
+
+    @Specialization
     public void display(final String argument) {
         Context.get(this).getOutput().println(argument);
     }