+++ language/src/main/java/i4gl/nodes/variables/write/WriteLocalVariable.java
@@ -12,6 +12,10 @@ import com.oracle.truffle.api.instrumentation.Tag;
 import i4gl.nodes.expression.ExpressionNode;
 import i4gl.nodes.statement.StatementNode;
 import i4gl.runtime.types.BaseType;
+import i4gl.runtime.types.compound.CharType;
+import i4gl.runtime.types.compound.VarcharType;
+import i4gl.runtime.values.Char;
+import i4gl.runtime.values.Varchar;
 
 @NodeChild(value = "valueNode", type = ExpressionNode.class)
 @NodeField(name = "slot", type = FrameSlot.class)
@@ -67,7 +71,31 @@ public abstract class WriteLocalVariable extends StatementNode {
         frame.setDouble(getSlot(), value);
     }
 
-    @Specialization(replaces = { "writeInt", "writeLong", "writeFloat", "writeDouble" })
+    protected boolean isChar() {
+        return getType() instanceof CharType;
+    }
+
+    @Specialization(guards = "isChar()")
+    protected void writeChar(VirtualFrame frame, String string) {
+        frame.getFrameDescriptor().setFrameSlotKind(getSlot(), FrameSlotKind.Object);
+        Char value = (Char) getType().getDefaultValue();
+        value.assignString(string);
+        frame.setObject(getSlot(), value);
+    }
+
+    protected boolean isVarchar() {
+        return getType() instanceof VarcharType;
+    }
+
+    @Specialization(guards = "isVarchar()")
+    protected void writeVarchar(VirtualFrame frame, String string) {
+        frame.getFrameDescriptor().setFrameSlotKind(getSlot(), FrameSlotKind.Object);
+        Varchar value = (Varchar) getType().getDefaultValue();
+        value.assignString(string);
+        frame.setObject(getSlot(), value);
+    }
+
+    @Specialization(replaces = { "writeInt", "writeLong", "writeFloat", "writeDouble", "writeChar", "writeVarchar" })
     protected void write(VirtualFrame frame, Object value) {
         frame.getFrameDescriptor().setFrameSlotKind(getSlot(), FrameSlotKind.Object);
         frame.setObject(getSlot(), value);