+++ language/src/main/java/i4gl/nodes/variables/write/WriteRecordFieldNode.java
@@ -9,7 +9,11 @@ import com.oracle.truffle.api.instrumentation.Tag;
 import i4gl.nodes.expression.ExpressionNode;
 import i4gl.nodes.statement.StatementNode;
 import i4gl.runtime.types.BaseType;
+import i4gl.runtime.types.compound.CharType;
+import i4gl.runtime.types.compound.VarcharType;
+import i4gl.runtime.values.Char;
 import i4gl.runtime.values.Record;
+import i4gl.runtime.values.Varchar;
 
 @NodeChild(value = "recordNode", type = ExpressionNode.class)
 @NodeChild(value = "valueNode", type = ExpressionNode.class)
@@ -21,13 +25,14 @@ public abstract class WriteRecordFieldNode extends StatementNode {
 
     // TODO: Check correct type at runtime
     protected abstract BaseType getFieldType();
-/*
-    @Specialization(guards = "record.isChar(getIdentifier())")
-    protected void writeChar1(Record record, char value) {
-        throw new UnexpectedRuntimeException();
-        // record.setChar(getIdentifier(), value);
-    }
-*/
+
+    /*
+     * @Specialization(guards = "record.isChar(getIdentifier())")
+     * protected void writeChar1(Record record, char value) {
+     * throw new UnexpectedRuntimeException();
+     * // record.setChar(getIdentifier(), value);
+     * }
+     */
     @Specialization(guards = "record.isSmallInt(getIdentifier())")
     protected void writeSmallInt(Record record, short value) {
         record.setSmallInt(getIdentifier(), value);
@@ -53,7 +58,29 @@ public abstract class WriteRecordFieldNode extends StatementNode {
         record.setFloat(getIdentifier(), value);
     }
 
-    @Specialization(replaces = { /*"writeChar1", */"writeSmallInt", "writeInt", "writeBigInt", "writeSmallFloat",
+    protected boolean isChar(Record record) {
+        return record.getRecordType().getFieldType(getIdentifier()) instanceof CharType;
+    }
+
+    @Specialization(guards = "isChar(record)")
+    protected void writeChar(Record record, String string) {
+        Char value = (Char) getFieldType().getDefaultValue();
+        value.assignString(string);
+        record.setObject(getIdentifier(), value);
+    }
+
+    protected boolean isVarchar(Record record) {
+        return record.getRecordType().getFieldType(getIdentifier()) instanceof VarcharType;
+    }
+
+    @Specialization(guards = "isVarchar(record)")
+    protected void writeVarchar(Record record, String string) {
+        Varchar value = (Varchar) getFieldType().getDefaultValue();
+        value.assignString(string);
+        record.setObject(getIdentifier(), value);
+    }
+
+    @Specialization(replaces = { /* "writeChar1", */"writeSmallInt", "writeInt", "writeBigInt", "writeSmallFloat",
             "writeFloat" })
     protected void write(Record record, Object value) {
         record.setObject(getIdentifier(), value);