+++ language/src/main/java/i4gl/runtime/types/compound/VarcharType.java
@@ -3,16 +3,16 @@ package i4gl.runtime.types.compound;
 import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.interop.InteropLibrary;
 
+import i4gl.runtime.types.primitive.ArrayType;
 import i4gl.runtime.values.Varchar;
 
 /**
  * Type descriptor representing the string type.
  */
-public class VarcharType extends TextType {
-    private final int size;
+public class VarcharType extends ArrayType {
 
     public VarcharType(int size) {
-        this.size = size;
+        super(size, Char1Type.SINGLETON);
     }
 
     @Override